backend/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
playwright>=1.49.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
pydantic>=2.10.0
httpx>=0.28.0
